+++ src/system.rs
@@ -1,5 +1,31 @@
-use sha3::{Digest, Sha3_512};
+/*
+ * This is a list containing all the simple characters that can be used to
+ * 1. Name users and groups
+ * 2. Name system environmental variables
+ * 3. Name a host on a network
+ */
+const SIMPLE_CHARACTERS:[char;64] = [
+    'a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k', 'l', 'm', 'n', 'i', 'j', 'k', 'l', 'm', 'n', 'o', 'p', 'q', 'r', 's', 't', 'u', 'v', 'w', 'x', 'y', 'z',
+    'A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J', 'K', 'L', 'M', 'N', 'I', 'J', 'K', 'L', 'M', 'N', 'O', 'P', 'Q', 'R', 'S', 'T', 'U', 'V', 'W', 'X', 'Y', 'Z'
+];
+// These characters are allowed too, but not in the beginning
+const SIMPLE_CHARACTERS_NOT_STARTING:[char;13] = ['0', '1', '2', '3', '4', '5', '6', '7', '8', '9', '.', '-', '_'];
 
+//
+pub fn check_simple_characters_compliance<S: AsRef<str>>(input: S) -> Result<(), &'static str> {
+    let input = input.as_ref();
+    for (i, c) in input.char_indices().enumerate() {
+        if i == 0 && SIMPLE_CHARACTERS_NOT_STARTING.contains(&c.1) {
+            return Err("This text can't start with a number, period, hyphen nor the underscore");
+        }
+        if !SIMPLE_CHARACTERS.contains(&c.1) || !SIMPLE_CHARACTERS_NOT_STARTING.contains(&c.1) {
+            return Err("This text may contain simple characters only");
+        }
+    }
+    Ok(())
+}
+
+use sha3::{Digest, Sha3_512};
 // Check which user/group is currently running
 pub fn current_user_real() -> Result<u32, &'static str> {
     extern "C" {